SIGTERM handler to fix cleanup in Jenkins
Jenkins' GitHub Pull Request Builder plugin sends SIGTERM to our script
when a PR branch is force-pushed, to cancel the current job and run a new
one. However, SIGTERM won't call AgentControl's destructor (__del__), which
returns the loaned node back to the pool.

To avoid this issue, let's move the cleanup code to the `finally`
section of the `try - except` block and raise a custom exception
when SIGTERM is caught.