vagrant: let's enable TEST-01-BASIC under clang+ASan
This needs some special tweaking in both CentOS CI and systemd
testsuite. First of all, systemd must be compiled with -shared-libasan,
so the ASan library is linked dynamically instead of statically. This
allows us pre-load it for processes, which makes use of the libsystemd
DSO, but aren't compiled with ASan themselves.

As the clang's ASan DSO is not in a standard library path (i.e. it's not
detected by ldconfig automatically), we need to play around with
LD_LIBRARY_PATH and lconfig config files to make sure each binary with
ASan can load the DSO as expected.

See the systemd counterpart PR: systemd/systemd#12663