agent: correctly collect meson logs
Originally, the meson logs were collected in the final EXIT handler with
the rest of the logs. However, as we use a relative path for them and we
change the CWD multiple times throughout the bootstrap process, the
relative path may become invalid (and it indeed does) during error
handling. Let's register a separate EXIT handler for this with an
absolute path, so we get the meson logs in all cases