vagrant: do a sanity boot check before reboot
In some cases the just compiled revision may end up being unbootable,
and since we don't have access to the serial console of the VM, it's
more or less undebuggable. Let's do a simple sanity check before the
reboot itself (as we already do in the CentOS 7 job) to make
the debugging easier.