vagrant: try to get a journal dump when ssh unexpectedly dies
In rare cases the `vagrant ssh` command dies unexpectedly (so far only
during the systemd-networkd testsuite) with EC 255. Let's add some
'hooks' to make this debuggable.

This, of course, works only if the sudden connection loss was just
intermittent. If the network gets completely broken, this won't work as
well.