vagrant: don't install tests in sanitizers jobs
We don't use them anyway and they break the sanitizers jobs during
soname bumps (libsystemd.so).

Reasoning: when we run tests directly from the build directory, the
libsystemd.so is correctly resolved thanks to RPATH being set to $ORIGIN
(i.e. runtime directory of the respective binary), so instead of a
system libsystemd DSO we use the one we just built.

However, in case of installed tests (-Dinstall-tests=true), these tests
are under /lib/systemd/tests, but the *correct* libsystemd DSO is not in
the binary's RPATH (as it's installed usually under /lib64), which
results in the system one being used. And as we don't run `ninja -C build
install`, the system libsystemd DSO has a different version than the
built one, which leads to nasty errors.

Everything above applies to the integration tests (test/TEST-??-*).

See: https://github.com/systemd/systemd/pull/13886#issuecomment-548788783