agent: skip TEST-13-NSPAWN-SMOKE subtests using cgroups v2
Even though RHEL 8 supports cgroups v2, we can't run systemd-nspawn with
unified cgroup hierarchy on it, as it still uses the legacy hierarchy.
As the auto-detection checks only for cgroups v2 support and not which
hierarchy is currently used on the host system, let's override the
auto-detection completely and skip tests using the unified hierarchy.