vagrant: configure nfsd explicitly
As Vagrant uses NFS for its shared folders, we need a working nfsd on
the host machine. So far we always left the responsibility on the
provisioner, but experience showed the nfsd is not always up when we
need it. Let's enable & start it explicitly to make sure we have a
working nfsd before playing around with Vagrant.