vagrant: check and possibly load the nfs & nfsd kernel modules
Beef up the nfsd check once again, as the race during Vagrant startup is
still there even though the nfsd seems to be running fine:

```
It appears your machine doesn't support NFS, or there is not an
adapter to enable NFS on this machine for Vagrant. Please verify
that `nfsd` is installed on your machine, and try again. If you're
on Windows, NFS isn't supported. If the problem persists, please
contact Vagrant support.
```