vagrant: move compilation away from NFS
The NFS volume we mount into the Vagrant VM with the systemd repository
has been giving me a headache for the past several months. Basically,
some of the Vagrant jobs (~5% of them) die due to the following issue:

```
/usr/bin/ld: error: /build/build/src/udev/libudev.so.1: file too short
collect2: error: ld returned 1 exit status
```

Apparently, this has to do something with desynchronized file timestamps
between the host and the guest, but even after installing and
configuring an NTP server on both sides the issue never went completely
away (see #201).

Let's just do the heavy lifting (compilation) on the internal VM storage
and see how it fares.

Fixes: #199