common: fix getting gdb traces when the build is not installed locally
When running `coredumpctl debug` on journal files got from the
integration tests (particularly QEMU ones), we get useless results if
the build is not installed locally (i.e. without running `ninja
install`), as coredumpctl pulls in "incorrect" binaries - e.g
coredumpctl pulls in `/usr/bin/systemctl` it got from the
`COREDUMP_FILENAME=` journal field, but that file is different from the
one in the $BUILD_DIR, resulting usually in a useless trace.

Let's fix this by attempting to locate the binary from the
`COREDUMP_FILENAME=` field in the local $BUILD_DIR (if set) and if we
succeed, override the gdb command to load the just located binary before
running `bt full`.

Fixes: #237