vagrant: configure the NFS grace period
Yet another attempt to mitigate the still-present NFS issue:

```
It appears your machine doesn't support NFS, or there is not an
adapter to enable NFS on this machine for Vagrant. Please verify
that `nfsd` is installed on your machine, and try again. If you're
on Windows, NFS isn't supported. If the problem persists, please
contact Vagrant support.
```

when starting VM. The rationale is that NFS server has a 90s grace
period (by default) after start during which it refuses any client
connections. Let's drop it to 10s (minimum valid value), since we're
in charge of all NFS clients in this case, and wait for the grace
period to pass, so we can be sure the NFS server is indeed ready
before trying to bring up a VM.