Revert "vagrant: halve the number of concurrent tasks"
This reverts commit 8779abe8f079f9a387564f72341e4fc3ad3f0e67.

After some debugging I noticed that the issue is still present, even
when the number of concurrent jobs is halved. However, when trying to
get deeper, the issue completely disappeared and I failed to reproduce
it in more than 20 consecutive runs.

In lights of this let's drop the workaround, since it doesn't work, and
wait if the issue reappears to dig into it even further.