vagrant: test the updated image before using it in production
So far we updated the Vagrant image every couple of days, gave it a
quick boot test to see if it's bootable, and then pushed it directly
into production. However, time has shown that this is actually not the
smartest idea, because there can be some interface changes or even bugs
introduced in some of the updated images, which completely breaks CI for
a while.

To avoid such scenarios, let's suffix the updated image with '-new', run
it through our usual pipelines, and if it succeeds drop the
'-new' suffix so the standard pipeline can use it without worrying it
may not work.