agent: reuse the corrected time & date after reboot
The corrected time & date doesn't persist across reboots, which means
that the machine will use the borked RTC time on the next boot. In that
case chrony attempts to fix it, but sometimes it does it too late in the
test phase. When this happens after some of the test images were
generated, they contain files with timestamps from the future, which
confuses the systemd unit cache after the time gets corrected, leading
to many unexpected (and bogus) fails because of missing unit files.

Related: https://github.com/systemd/systemd-centos-ci/issues/251#issuecomment-648180867