vagrant: explicitly compile krb5 with the builtin crypto
The reason why this workaround is even needed in the first place is that
the distro libk5crypto.so is built with the OpenSSL crypto backend
(--with-crypto-impl=openssl), which causes undefined symbol errors, as
the Vagrant's embedded OpenSSL is older than the distro one.

To workaround this, we compile a local version of libk5crypto.so using
the builtin crypto implementation (--with-crypto-impl=builtin). Let's
make this fact explicit and add some nice documentation, so my future
dumb me doesn't have to spend an hour debugging this mess (again).