utils: use curl instead of wget
As of 2021-09-30 one of the Let's Encrypt's root certificates expired,
which in combination with CentOS 7's OpenSSL causes verification fails
in tools using OpenSSL (like wget). Let's use curl instead, which uses
NSS that doesn't suffer from this issue.

Another option is to upgrade the `ca-certificates` package, but since we
don't have root in the OCP container, this is not possible.

See:
    * https://letsencrypt.org/docs/dst-root-ca-x3-expiration-september-2021/
    * https://access.redhat.com/articles/6338021