agent-control: make the allocation wait-loop even more aggressive
There's an unfortunate bug in the Kubernetes Jenkins module, when it
sometimes spawns more concurrent pods than it should, which leads to
starving jobs that are actively waiting for a free session. Let's reduce
the sleep interval from 2 minutes to 5 seconds, so we can snatch the
free session before Jenkins provisions a new pod.