agent(upstream): prefer nspawn when running tests
As the situation with overloaded AWS regions got quite unbearable and
most of the C8S jobs keep failing with timeouts, let's resort to running most
of the tests in nspawn only and use qemu only when absolutely necessary.
This is something that Ubuntu CI already does for quite a long time, so
the test sutie is already fine-tuned to use qemu only when we really have to.

We shouldn't lose any coverage (or barely any) as we still run both
parts in the Arch Linux job which doesn't suffer from this issue as it
runs on a baremetal machine, not a VM.

This issue is, unfortunately, out of my reach, as that's how the machine
pool in AWS is configured, see [0].

[0] https://lists.centos.org/pipermail/ci-users/2022-October/004618.html