agent: run only nspawn tests in the C8S job on PRs
Since the issues with overloaded AWS regions are still there and are
unlikely to disappear in a foreseeable future, let's give up on the QEMU
stuff and run only tests that run in nspawn. We'll lose some coverage on
C8S (mainly from TEST-64), but it's still better than not running
anything.

To make up for this, I'll set up a cron job that'll run the whole test
suite at least once a day in the "non-busy" hours.