agent-control: unify node cleanup paths
Also, let's try extra hard when cleaning up a session on a signal (i.e.
when a Jenkins job is aborted). Once we get the first SIGTERM, there's a
short window we have for the cleanup before we get SIGKILLed, so let's
get straight down to it and just call sys.exit(). This should
immediately trigger the class destructor, which in turn does the actual
node cleanup.

The current exception approach didn't work in all cases, so let's see if
this help a bit.