Ignore coredumps generated by unit test
There's quite a number of intentional crashes in unit tests and
filtering them out is PITA, especially when it comes to crashes that
happen during "double-fork" (test-execute), since their executable is
then set to systemd-executor, which we don't want to "blanket-ignore".

Instead, let's check for any coredumps before running unit tests, so we
get the boot sequence covered, and then instruct the coredump-collecting
machinery to look for coredumps that happened only after the unit tests.