Allow sanitized binaries to dump a core
This should make debugging easier if the binary crashes outside of the
ASAn/UBSan stuff, since in that case we have neither the coredump nor a
stack trace. It should be enough to set disable_coredumps=0, since
use_madv_dontdump is enabled by default, but let's play it safe and
avoid possibly dumping a 16+ TB core in case something flips that bit
off.