common: drop setting `debuginfod`
gdb should do this automagically [0], so let's drop this (as the option
is apparently not in C9S's gdb even though it supports debuginfod just
fine). However, let's explicitly source /etc/profile.d/debuginfod.sh if
it exists, since in certain cases we might install gdb and not restart
the shell, so we never get the $DEBUGINFOD_URLS env variable, and end up
without the debuginfod stuff.

This effectively reverts 0c17b85.

[0] https://wiki.archlinux.org/title/Debuginfod

sq