agent(rhel8): move the C8S jobs to C9S
Since C8S is EOL, but we still support RHEL 8, let's run the integration
tests on C9S instead. This won't exactly represent the RHEL 8
environment, but it should be enough to catch early regressions - we
still run the same tests internally on RHEL 8 for the "proper"
verification.

To make the tests work on C9S, some changes are necessary:
  - use dbus-daemon instead of dbus-broker (since RHEL 8's systemd test
    suite is not compatible with dbus-broker)
  - skip parse-hwdb and test-execute, and they're not compatible with
    RHEL 9's package set (we still run them in other CIs anyway)
  - explicitly whitelist all known syscalls for nspawn containers, as
    systemd is missing some syscalls from RHEL 9 from its default
    whitelists (but it thankfully knows about them, so @known set works
    here)
  - don't build with --werror, as RHEL 9's gcc has some additional
    warnings enable for which we don't have the respective patches in
    RHEL 8's systemd; not sure (yet) how feasible it would be to
    backported them to RHEL 8